Complete the sample code of RTMPInjection and RTMPStreaming.
