+++ Android/APIExample/app/src/main/java/io/agora/api/example/examples/live_broadcasting/RTMPStreaming.java
@@ -21,10 +21,11 @@ import com.yanzhenjie.permission.runtime.Permission;
 import io.agora.api.example.R;
 import io.agora.api.example.annotation.Example;
 import io.agora.api.example.common.BaseFragment;
+import io.agora.api.example.examples.basic_video_audio.VideoShow;
 import io.agora.rtc.Constants;
 import io.agora.rtc.IRtcEngineEventHandler;
 import io.agora.rtc.RtcEngine;
-import io.agora.rtc.live.LiveInjectStreamConfig;
+import io.agora.rtc.live.LiveTranscoding;
 import io.agora.rtc.video.VideoCanvas;
 import io.agora.rtc.video.VideoEncoderConfiguration;
 
@@ -34,34 +35,32 @@ import static io.agora.rtc.video.VideoEncoderConfiguration.ORIENTATION_MODE.ORIE
 import static io.agora.rtc.video.VideoEncoderConfiguration.STANDARD_BITRATE;
 import static io.agora.rtc.video.VideoEncoderConfiguration.VD_640x360;
 
-/**
- * This example demonstrates how to pull flow from an external address.
+/**This example demonstrates how to push a stream to an external address.
  *
  * Important:
  *          Users who push and pull streams cannot be in one channel,
- *          otherwise unexpected errors will occur.
- */
+ *          otherwise unexpected errors will occur.*/
 @Example(
         group = "Live BROADCASTING",
-        name = "RTMP Injection",
-        actionId = R.id.action_mainFragment_to_RTMPInjection
+        name = "RTMP Streaming",
+        actionId = R.id.action_mainFragment_to_RTMPStreaming
 )
-public class RTMPInjection extends BaseFragment implements View.OnClickListener
+public class RTMPStreaming extends BaseFragment implements View.OnClickListener
 {
-    private static final String TAG = RTMPInjection.class.getSimpleName();
+    private static final String TAG = RTMPStreaming.class.getSimpleName();
 
     private FrameLayout fl_local, fl_remote;
     private EditText et_url, et_channel;
-    private Button join, inject;
+    private Button join, publish;
     private RtcEngine engine;
     private int myUid;
-    private boolean joined = false, injecting = false;
+    private boolean joined = false, publishing = false;
 
     @Nullable
     @Override
     public View onCreateView(@NonNull LayoutInflater inflater, @Nullable ViewGroup container, @Nullable Bundle savedInstanceState)
     {
-        View view = inflater.inflate(R.layout.fragment_rtmp_injection, container, false);
+        View view = inflater.inflate(R.layout.fragment_rtmp_streaming, container, false);
         return view;
     }
 
@@ -75,8 +74,8 @@ public class RTMPInjection extends BaseFragment implements View.OnClickListener
         et_url = view.findViewById(R.id.et_url);
         join = view.findViewById(R.id.btn_join);
         join.setOnClickListener(this);
-        inject = view.findViewById(R.id.btn_inject);
-        inject.setOnClickListener(this);
+        publish = view.findViewById(R.id.btn_publish);
+        publish.setOnClickListener(this);
     }
 
     @Override
@@ -151,21 +150,21 @@ public class RTMPInjection extends BaseFragment implements View.OnClickListener
                 engine.leaveChannel();
                 joined = false;
                 join.setText(getString(R.string.join));
-                injecting = false;
-                inject.setEnabled(false);
-                inject.setText(getString(R.string.inject));
+                publishing = false;
+                publish.setEnabled(false);
+                publish.setText(getString(R.string.publish));
             }
         }
-        else if (v.getId() == R.id.btn_inject)
+        else if (v.getId() == R.id.btn_publish)
         {
             /**Ensure that the user joins a channel before calling this method.*/
-            if(joined && !injecting)
+            if(joined && !publishing)
             {
-                startInjection();
+                startPublish();
             }
-            else if(joined && injecting)
+            else if(joined && publishing)
             {
-                stopInjection();
+                stopPublish();
             }
         }
     }
@@ -236,60 +235,89 @@ public class RTMPInjection extends BaseFragment implements View.OnClickListener
         join.setEnabled(false);
     }
 
-    private void startInjection()
+    private void startPublish()
     {
-        /**Configuration of the imported live broadcast voice or video stream.
-         * See <a href="https://docs.agora.io/en/Video/API%20Reference/java/classio_1_1agora_1_1rtc_1_1live_1_1_live_inject_stream_config.html"></a>*/
-        LiveInjectStreamConfig config = new LiveInjectStreamConfig();
-        /**Injects an online media stream to a live broadcast.
-         * If this method call is successful, the server pulls the voice or video stream and injects
-         * it into a live channel. This is applicable to scenarios where all audience members in the
-         * channel can watch a live show and interact with each other.
-         * The addInjectStreamUrl method call triggers the following callbacks:
-         *   The local client:
-         *      onStreamInjectedStatus, with the state of the injecting the online stream.
-         *      onUserJoined(uid: 666), if the method call is successful and the online media stream
-         *          is injected into the channel.
-         *   The remote client:
-         *      onUserJoined(uid: 666), if the method call is successful and the online media stream
-         *          is injected into the channel.
-         * @param url The URL address to be added to the ongoing live broadcast. Valid protocols are RTMP, HLS, and HTTP-FLV.
-         *              Supported FLV audio codec type: AAC.
-         *              Supported FLV video codec type: H264(AVC).
-         * @param config The LiveInjectStreamConfig object which contains the configuration information
-         *               for the added voice or video stream.
+        /**LiveTranscoding: A class for managing user-specific CDN live audio/video transcoding settings.
+         * See <a href="https://docs.agora.io/en/Video/API%20Reference/java/classio_1_1agora_1_1rtc_1_1live_1_1_live_transcoding.html"></a>*/
+        LiveTranscoding transcoding = new LiveTranscoding();
+        /**The transcodingUser class which defines the video properties of the user displaying the
+         * video in the CDN live. Agora supports a maximum of 17 transcoding users in a CDN live streaming channel.
+         * See <a href="https://docs.agora.io/en/Video/API%20Reference/java/classio_1_1agora_1_1rtc_1_1live_1_1_live_transcoding_1_1_transcoding_user.html"></a>*/
+        LiveTranscoding.TranscodingUser transcodingUser = new LiveTranscoding.TranscodingUser();
+        transcodingUser.width = transcoding.width;
+        transcodingUser.height = transcoding.height;
+        transcodingUser.uid = myUid;
+        /**Adds a user displaying the video in CDN live.
+         * @return
+         *  0: Success.
+         *  <0: Failure.*/
+        int ret = transcoding.addUser(transcodingUser);
+        /**Sets the video layout and audio settings for CDN live.
+         * The SDK triggers the onTranscodingUpdated callback when you call this method to update
+         * the LiveTranscodingclass. If you call this method to set the LiveTranscoding class for
+         * the first time, the SDK does not trigger the onTranscodingUpdated callback.
+         * @param transcoding Sets the CDN live audio/video transcoding settings See
+         *   <a href="https://docs.agora.io/en/Video/API%20Reference/java/classio_1_1agora_1_1rtc_1_1live_1_1_live_transcoding.html"></a>
+         * @return
+         *   0: Success.
+         *   <0: Failure.
+         * PS:
+         *   This method applies to Live Broadcast only.
+         *   Ensure that you enable the RTMP Converter service before using this function. See
+         *      Prerequisites in Push Streams to CDN.
+         *   Ensure that you call the setClientRole method and set the user role as the host.
+         *   Ensure that you call the setLiveTranscoding method before calling the addPublishStreamUrl method.*/
+        engine.setLiveTranscoding(transcoding);
+        /**Publishes the local stream to the CDN.
+         * The addPublishStreamUrl method call triggers the onRtmpStreamingStateChanged callback on
+         * the local client to report the state of adding a local stream to the CDN.
+         * @param url The CDN streaming URL in the RTMP format. The maximum length of this parameter
+         *            is 1024 bytes. The URL address must not contain special characters, such as
+         *            Chinese language characters.
+         * @param transcodingEnabled Sets whether transcoding is enabled/disabled. If you set this
+         *                           parameter as true, ensure that you call the setLiveTranscoding
+         *                           method before this method.
+         *                              true: Enable transcoding. To transcode the audio or video
+         *                                 streams when publishing them to CDN live, often used for
+         *                                 combining the audio and video streams of multiple hosts in CDN live.
+         *                              false: Disable transcoding.
          * @return
          *   0: Success.
          *   < 0: Failure.
-         *      ERR_INVALID_ARGUMENT(2): The injected URL does not exist. Call this method again to
-         *          inject the stream and ensure that the URL is valid.
-         *      ERR_NOT_READY(3): The user is not in the channel.
-         *      ERR_NOT_SUPPORTED(4): The channel profile is not Live Broadcast. Call the setChannelProfile
-         *          method and set the channel profile to Live Broadcast before calling this method.
-         *      ERR_NOT_INITIALIZED(7): The SDK is not initialized. Ensure that the RtcEngine object
-         *          is initialized before using this method.
+         *      ERR_INVALID_ARGUMENT(2): Invalid parameter, usually because the URL address is null or the string length is 0.
+         *      ERR_NOT_INITIALIZED(7): You have not initialized RtcEngine when publishing the stream.
          * PS:
-         *   This method applies to the Live-Broadcast profile only.
          *   Ensure that you enable the RTMP Converter service before using this function. See
          *      Prerequisites in Push Streams to CDN.
-         *   You can inject only one media stream into the channel at the same time.*/
-        engine.addInjectStreamUrl(et_url.getText().toString(), config);
+         *   This method applies to Live Broadcast only.
+         *   Ensure that the user joins a channel before calling this method.
+         *   This method adds only one stream HTTP/HTTPS URL address each time it is called.*/
+        int code = engine.addPublishStreamUrl(et_url.getText().toString(), true);
+        /**Prevent repeated entry*/
+        publish.setEnabled(false);
     }
 
-    private void stopInjection()
+    private void stopPublish()
     {
-        injecting = false;
-        inject.setEnabled(true);
-        inject.setText(getString(R.string.inject));
-        /**Removes the injected online media stream from a live broadcast.
-         * This method removes the URL address (added by addInjectStreamUrl) from a live broadcast.
-         * If this method call is successful, the SDK triggers the onUserOffline callback and returns
-         * a stream uid of 666.
-         * @param url HTTP/HTTPS URL address of the added stream to be removed.
+        publishing = false;
+        publish.setEnabled(true);
+        publish.setText(getString(R.string.publish));
+        /**Removes an RTMP stream from the CDN.
+         * This method removes the RTMP URL address (added by addPublishStreamUrl) from a CDN live
+         * stream. The SDK reports the result of this method call in the onRtmpStreamingStateChanged callback.
+         * @param url The RTMP URL address to be removed. The maximum length of this parameter is
+         *            1024 bytes. The URL address must not contain special characters, such as
+         *            Chinese language characters.
          * @return
          *   0: Success.
-         *   < 0: Failure.*/
-        int ret = engine.removeInjectStreamUrl(et_url.getText().toString());
+         *   <0: Failure.
+         * PS:
+         *   Ensure that you enable the RTMP Converter service before using this function. See
+         *      Prerequisites in Push Streams to CDN.
+         *   Ensure that the user joins a channel before calling this method.
+         *   This method applies to Live Broadcast only.
+         *   This method removes only one stream RTMP URL address each time it is called.*/
+        int ret = engine.removePublishStreamUrl(et_url.getText().toString());
     }
 
     /**
@@ -346,8 +374,8 @@ public class RTMPInjection extends BaseFragment implements View.OnClickListener
                 {
                     join.setEnabled(true);
                     join.setText(getString(R.string.leave));
-                    inject.setEnabled(true);
-                    inject.setText(getString(R.string.inject));
+                    publish.setEnabled(true);
+                    publish.setText(getString(R.string.publish));
                 }
             });
         }
@@ -435,37 +463,68 @@ public class RTMPInjection extends BaseFragment implements View.OnClickListener
             Log.i(TAG, "onRemoteVideoStateChanged->" + uid + ", state->" + state + ", reason->" + reason);
         }
 
-        /**Reports the status of injecting the online media stream.
-         * @param url The URL address of the externally injected stream.
-         * @param uid User ID.
-         * @param status
-         *   INJECT_STREAM_STATUS_START_SUCCESS(0): The external video stream imports successfully.
-         *   INJECT_STREAM_STATUS_START_ALREADY_EXIST(1): The external video stream already exists.
-         *   INJECT_STREAM_STATUS_START_UNAUTHORIZED(2): The external video stream import is unauthorized.
-         *   INJECT_STREAM_STATUS_START_TIMEDOUT(3): Timeout when importing the external video stream.
-         *   INJECT_STREAM_STATUS_START_FAILED(4): The external video stream fails to import.
-         *   INJECT_STREAM_STATUS_STOP_SUCCESS(5): The external video stream stops importing successfully.
-         *   INJECT_STREAM_STATUS_STOP_NOT_FOUND(6): No external video stream is found.
-         *   INJECT_STREAM_STATUS_STOP_UNAUTHORIZED(7): The external video stream stops from being unauthorized.
-         *   INJECT_STREAM_STATUS_STOP_TIMEDOUT(8): Timeout when stopping the import of the external video stream.
-         *   INJECT_STREAM_STATUS_STOP_FAILED(9): Fails to stop importing the external video stream.
-         *   INJECT_STREAM_STATUS_BROKEN(10): The external video stream import is interrupted.*/
+        /**Since v2.4.1
+         * Occurs when the state of the RTMP streaming changes.
+         * This callback indicates the state of the RTMP streaming. When exceptions occur, you can
+         * troubleshoot issues by referring to the detailed error descriptions in the errCode parameter.
+         * @param url The RTMP URL address.
+         * @param state The RTMP streaming state:
+         *   RTMP_STREAM_PUBLISH_STATE_IDLE(0): The RTMP streaming has not started or has ended.
+         *              This state is also triggered after you remove an RTMP address from the CDN
+         *              by calling removePublishStreamUrl.
+         *   RTMP_STREAM_PUBLISH_STATE_CONNECTING(1): The SDK is connecting to Agora streaming server
+         *              and the RTMP server. This state is triggered after you call the addPublishStreamUrl method.
+         *   RTMP_STREAM_PUBLISH_STATE_RUNNING(2): The RTMP streaming publishes. The SDK successfully
+         *              publishes the RTMP streaming and returns this state.
+         *   RTMP_STREAM_PUBLISH_STATE_RECOVERING(3): The RTMP streaming is recovering. When exceptions
+         *              occur to the CDN, or the streaming is interrupted, the SDK attempts to resume
+         *              RTMP streaming and returns this state.
+         *                1:If the SDK successfully resumes the streaming, RTMP_STREAM_PUBLISH_STATE_RUNNING(2)
+         *                    returns.
+         *                2:If the streaming does not resume within 60 seconds or server errors occur,
+         *                    RTMP_STREAM_PUBLISH_STATE_FAILURE(4) returns. You can also reconnect to the
+         *                    server by calling the removePublishStreamUrl and addPublishStreamUrl methods.
+         *   RTMP_STREAM_PUBLISH_STATE_FAILURE(4): The RTMP streaming fails. See the errCode parameter
+         *              for the detailed error information. You can also call the addPublishStreamUrl
+         *              method to publish the RTMP streaming again.
+         * @param errCode The detailed error information for streaming:
+         *   RTMP_STREAM_PUBLISH_ERROR_OK(0): The RTMP streaming publishes successfully.
+         *   RTMP_STREAM_PUBLISH_ERROR_INVALID_ARGUMEN(1): Invalid argument used. If, for example,
+         *                you do not call the setLiveTranscoding method to configure the LiveTranscoding
+         *                parameters before calling the addPublishStreamUrl method, the SDK returns
+         *                this error. Check whether you set the parameters in the setLiveTranscoding method properly.
+         *   RTMP_STREAM_PUBLISH_ERROR_ENCRYPTED_STREAM_NOT_ALLOWED(2): The RTMP streaming is
+         *                encrypted and cannot be published.
+         *   RTMP_STREAM_PUBLISH_ERROR_CONNECTION_TIMEOUT(3): Timeout for the RTMP streaming. Call
+         *                the addPublishStreamUrl method to publish the streaming again.
+         *   RTMP_STREAM_PUBLISH_ERROR_INTERNAL_SERVER_ERROR(4): An error occurs in Agora streaming
+         *                server. Call the addPublishStreamUrl method to publish the streaming again.
+         *   RTMP_STREAM_PUBLISH_ERROR_RTMP_SERVER_ERROR(5): An error occurs in the RTMP server.
+         *   RTMP_STREAM_PUBLISH_ERROR_TOO_OFTEN(6): The RTMP streaming publishes too frequently.
+         *   RTMP_STREAM_PUBLISH_ERROR_REACH_LIMIT(7): The host publishes more than 10 URLs. Delete
+         *                the unnecessary URLs before adding new ones.
+         *   RTMP_STREAM_PUBLISH_ERROR_NOT_AUTHORIZED(8): The host manipulates other hosts' URLs.
+         *                Check your app logic.
+         *   RTMP_STREAM_PUBLISH_ERROR_STREAM_NOT_FOUND(9): Agora server fails to find the RTMP
+         *                streaming.
+         *   RTMP_STREAM_PUBLISH_ERROR_FORMAT_NOT_SUPPORTED(10): The format of the RTMP streaming
+         *                URL is not supported. Check whether the URL format is correct.*/
         @Override
-        public void onStreamInjectedStatus(String url, int uid, int status)
+        public void onRtmpStreamingStateChanged(String url, int state, int errCode)
         {
-            super.onStreamInjectedStatus(url, uid, status);
-            Log.i(TAG, "onStreamInjectedStatus->" + url + ", uid->" + uid + ", status->" + status);
-            if(status == Constants.INJECT_STREAM_STATUS_START_SUCCESS)
+            super.onRtmpStreamingStateChanged(url, state, errCode);
+            Log.i(TAG, "onRtmpStreamingStateChanged->" + url + ", state->" + state + ", errCode->" + errCode);
+            if(state == Constants.RTMP_STREAM_PUBLISH_STATE_RUNNING)
             {
                 /**After confirming the successful push, make changes to the UI.*/
-                injecting = true;
+                publishing = true;
                 handler.post(new Runnable()
                 {
                     @Override
                     public void run()
                     {
-                        inject.setEnabled(true);
-                        inject.setText(getString(R.string.stopinject));
+                        publish.setEnabled(true);
+                        publish.setText(getString(R.string.stoppublish));
                     }
                 });
             }
@@ -486,14 +545,23 @@ public class RTMPInjection extends BaseFragment implements View.OnClickListener
             if (context == null) return;
             handler.post(() ->
             {
-                fl_remote.removeAllViews();
                 /**Display remote video stream*/
                 SurfaceView surfaceView = null;
-                // Create render view by RtcEngine
-                surfaceView = RtcEngine.CreateRendererView(context);
-                // Add to the remote container
-                fl_remote.addView(surfaceView, new FrameLayout.LayoutParams(ViewGroup.LayoutParams.MATCH_PARENT,
-                        ViewGroup.LayoutParams.MATCH_PARENT));
+                if (fl_remote.getChildCount() == 0)
+                {
+                    // Create render view by RtcEngine
+                    surfaceView = RtcEngine.CreateRendererView(context);
+                    // Add to the remote container
+                    fl_remote.addView(surfaceView, new FrameLayout.LayoutParams(ViewGroup.LayoutParams.MATCH_PARENT, ViewGroup.LayoutParams.MATCH_PARENT));
+                }
+                else
+                {
+                    View view = fl_remote.getChildAt(0);
+                    if (view instanceof SurfaceView)
+                    {
+                        surfaceView = (SurfaceView) view;
+                    }
+                }
 
                 // Setup remote video to render
                 engine.setupRemoteVideo(new VideoCanvas(surfaceView, RENDER_MODE_HIDDEN, uid));
