+++ Android/APIExample/app/src/main/res/layout/fragment_joinchannel_audio.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".examples.basic_video_audio.AudioOnly">
+    tools:context=".examples.basic_video_audio.JoinChannelAudio">
 
     <androidx.appcompat.widget.AppCompatButton
         android:id="@+id/btn_mute"
