+++ Android/APIExample/app/src/main/res/layout/fragment_joinchannel_video.xml
