--- Android/APIExample/app/src/main/res/layout/fragment_rtmp_publishing.xml
@@ -1,40 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:context=".examples.live_broadcasting.RTMPPublishing">
-
-    <include
-        layout="@layout/fragment_video_show"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:layout_above="@+id/ll_push" />
-
-    <LinearLayout
-        android:id="@+id/ll_push"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_alignParentBottom="true"
-        android:gravity="center_vertical"
-        android:orientation="horizontal">
-
-        <androidx.appcompat.widget.AppCompatEditText
-            android:id="@+id/et_url"
-            android:layout_width="0dp"
-            android:layout_height="wrap_content"
-            android:layout_weight="1"
-            android:hint="@string/url"
-            android:singleLine="true"
-            android:text="rtmp://vid-218.push.chinanetcenter.broadcastapp.agora.io/live/qwer" />
-
-        <androidx.appcompat.widget.AppCompatButton
-            android:id="@+id/btn_publish"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:singleLine="true"
-            android:text="@string/publish" />
-
-    </LinearLayout>
-
-</RelativeLayout>
