+++ Android/APIExample/app/src/main/res/navigation/nav_graph.xml
@@ -10,44 +10,36 @@
         android:label="@string/app_name"
         tools:layout="@layout/fragment_main">
         <action
-            android:id="@+id/action_mainFragment_to_joinChannel"
-            app:destination="@id/joinChannel" />
+            android:id="@+id/action_mainFragment_to_joinChannelAudio"
+            app:destination="@id/joinChannelAudio" />
         <action
-            android:id="@+id/action_mainFragment_to_audioOnly"
-            app:destination="@id/audioOnly" />
-        <action
-            android:id="@+id/action_mainFragment_to_videoShow"
-            app:destination="@id/videoShow" />
+            android:id="@+id/action_mainFragment_to_joinChannelVideo"
+            app:destination="@id/joinChannelVideo" />
         <action
             android:id="@+id/action_mainFragment_to_RTMPInjection"
             app:destination="@id/RTMPInjection" />
         <action
-            android:id="@+id/action_mainFragment_to_RTMPPublishing"
-            app:destination="@id/RTMPPublishing" />
+            android:id="@+id/action_mainFragment_to_RTMPStreaming"
+            app:destination="@id/RTMPStreaming" />
     </fragment>
     <fragment
-        android:id="@+id/joinChannel"
-        android:name="io.agora.api.example.examples.basic_video_audio.JoinChannel"
-        android:label="JoinChannel"
-        tools:layout="@layout/fragment_join_channel" />
+        android:id="@+id/joinChannelAudio"
+        android:name="io.agora.api.example.examples.basic_video_audio.JoinChannelAudio"
+        android:label="JoinChannelAudio"
+        tools:layout="@layout/fragment_joinchannel_audio" />
     <fragment
-        android:id="@+id/audioOnly"
-        android:name="io.agora.api.example.examples.basic_video_audio.AudioOnly"
-        android:label="AudioOnly"
-        tools:layout="@layout/fragment_audio_only" />
-    <fragment
-        android:id="@+id/videoShow"
-        android:name="io.agora.api.example.examples.basic_video_audio.VideoShow2"
-        android:label="VideoShow2"
-        tools:layout="@layout/fragment_video_show" />
+        android:id="@+id/joinChannelVideo"
+        android:name="io.agora.api.example.examples.basic_video_audio.JoinChannelVideo"
+        android:label="JoinChannelVideo"
+        tools:layout="@layout/fragment_joinchannel_video" />
     <fragment
         android:id="@+id/RTMPInjection"
         android:name="io.agora.api.example.examples.live_broadcasting.RTMPInjection"
         android:label="RTMPInjection"
         tools:layout="@layout/fragment_rtmp_injection" />
     <fragment
-        android:id="@+id/RTMPPublishing"
-        android:name="io.agora.api.example.examples.live_broadcasting.RTMPPublishing"
-        android:label="RTMPPublishing"
-        tools:layout="@layout/fragment_rtmp_publishing" />
+        android:id="@+id/RTMPStreaming"
+        android:name="io.agora.api.example.examples.live_broadcasting.RTMPStreaming"
+        android:label="RTMPStreaming"
+        tools:layout="@layout/fragment_rtmp_streaming" />
 </navigation>
