+++ Android/APIExample/app/src/main/res/values/string_configs.xml
@@ -2,6 +2,8 @@
 <resources>
 <!--    <string name="agora_app_id" translatable="false"><#YOUR APP ID#></string>-->
 <!--    <string name="agora_access_token" translatable="false"><#YOUR ACCESS TOKEN#></string>-->
-    <string name="agora_app_id" translatable="false">266819b117a042a9aa188a67b5daabfb</string>
-    <string name="agora_access_token" translatable="false">006266819b117a042a9aa188a67b5daabfbIACtkmziXmc2ZrBdrlU3rPftAlV4gy6IWEGlzGDhgIZMFBkiUXMAAAAAEADXbN9K3dLVXgEAAQDd0tVe</string>
+<!--    <string name="agora_app_id" translatable="false">266819b117a042a9aa188a67b5daabfb</string>-->
+<!--    <string name="agora_access_token" translatable="false">006266819b117a042a9aa188a67b5daabfbIADR+dUn7WoOJ6LIuT5E1A3H3K9FnHvYwZJgJJdCmFXTzhkiUXMAAAAAEADXbN9KGAXXXgEAAQAYBdde</string>-->
+    <string name="agora_app_id" translatable="false">aab8b8f5a8cd4469a63042fcfafe7063</string>
+    <string name="agora_access_token" translatable="false">111</string>
 </resources>
